Add files via upload
This is a program which uses a search term as input and returns places results using the google places API, and creates a csv file with a data table including information about the place like business name, contact, location info, type of business etc. You need to enter your own API key to use the program.